+++ src/quick-lint-js/fe/parse-class.cpp
@@ -303,9 +303,16 @@ void Parser::parse_and_visit_class_or_interface_member(
         // abstract f();
         // accessor myField = null;
         // async f() {}
+        // private f() {}
+        // public static field = 42;
+        // readonly field: number;
         case Token_Type::kw_abstract:
         case Token_Type::kw_accessor:
         case Token_Type::kw_async:
+        case Token_Type::kw_private:
+        case Token_Type::kw_protected:
+        case Token_Type::kw_public:
+        case Token_Type::kw_readonly:
           last_ident = p->peek().identifier_name();
           modifiers.push_back(Modifier{
               .span = p->peek().span(),
@@ -357,31 +364,8 @@ void Parser::parse_and_visit_class_or_interface_member(
           p->skip();
           continue;
 
-        // private f() {}
-        // public static field = 42;
-        case Token_Type::kw_private:
-        case Token_Type::kw_protected:
-        case Token_Type::kw_public:
-          last_ident = p->peek().identifier_name();
-          modifiers.push_back(Modifier{
-              .span = p->peek().span(),
-              .type = p->peek().type,
-          });
-          p->skip();
-          continue;
-
         // static f() {}
         case Token_Type::kw_static:
-          last_ident = p->peek().identifier_name();
-          modifiers.push_back(Modifier{
-              .span = p->peek().span(),
-              .type = p->peek().type,
-          });
-          p->skip();
-          continue;
-
-        // readonly field: number;
-        case Token_Type::kw_readonly:
           last_ident = p->peek().identifier_name();
           modifiers.push_back(Modifier{
               .span = p->peek().span(),